Fix variable type

+++ google-http-client/src/main/java/com/google/api/client/http/javanet/NetHttpResponse.java
@@ -188,7 +188,7 @@ final class NetHttpResponse extends LowLevelHttpResponse {
 
     @Override
     public long skip(long len) throws IOException {
-      int n = in.skip(len);
+      long n = in.skip(len);
       if (n == -1) {
         throwIfFalseEOF();
       } else {
